requirements/common.txt
# Core Stuff
# -------------------------------------
Django==2.1.5
djangorestframework==3.9.0

# Configuration
# -------------------------------------
python-dotenv==0.10.0

# Phone number related Stuff
# -------------------------------------
phonenumbers==8.10.2
django-phonenumber-field==2.1.0
twilio==6.21.0
